Add files via upload
With the help of the provided Python program, the user can easily calculate fuel transactions based on their preferences. The programme gives users the option to choose between Petrol and Diesel and whether they want to input values in rupees or litres, all while being guided by an intuitive set of prompts.
An algorithm that has been painstakingly designed executes the calculations precisely after receiving these choices. When choosing Petrol, the programme expertly multiplies the amount of money input to equal the amount of fuel. The programme competently converts fuel litres to corresponding rupee values for Diesel enthusiasts, demonstrating its versatility.
Whether it's the allure of a tank full of gasoline or the practical effectiveness of diesel, this programme places a priority on user convenience. It